fix: address the twelve findings from the multi-agent expert review
A 7-lens expert review with adversarial verification surfaced twelve confirmed
findings. All are fixed here.

Security:
  * Key leak (High). The 0x30 1W key-transfer frame that commit 2e2217b
    redacted survived verbatim in eight other places - seven raw-dump lines in
    docs/devices/velux/captures/README.md and one in velux-frame-analysis.md -
    each re-exposing the same masked key, from which the installation's system
    key is recoverable with the public transfer key. All redacted; the repo is
    verified free of the frame bytes. (The git-history caveat the original
    redaction recorded still stands.)

Correctness / robustness:
  * Null-deref crash (Medium). A node begun in 1W mode with accept_pairing_ set
    dereferenced a null auth_manager_ on an unauthenticated broadcast 0x31 - a
    single frame could crash every such node. is_pairing_frame() and
    handle_pairing_frame() now refuse to route pairing frames in 1W mode.
  * ESPHome cover (Medium). In 2W mode the OPEN/CLOSE/position path ignored the
    send result, so a command dropped while the session was still handshaking
    still animated the cover and published a wrong position. control() now only
    arms movement when the send succeeded, matching the tilt path.
  * Pairing timeout (Low). A stalled handshake never reset, diverting unrelated
    0x3C/0x32 frames from the MAC gate; pairing_state_ now expires after 5 s.
  * ChannelHopper (Low). Timestamp math is now done in uint32_t so the counter
    wraps at 2^32 on a 64-bit host too, matching Arduino micros() - previously
    the host and target diverged and the wrap path was never exercised.

Security hygiene / docs:
  * Key-capture log (Low). The ESPHome key_capture path printed the recovered
    key over the network logger; a loud security warning now precedes it.
  * Replay guard (Low). The receive-side ReplayGuard is RAM-only and resets on
    reboot - documented in the header and corrected in the threat table, which
    had over-credited "persisted rolling code" (that persistence is the
    transmit counter).
  * velux-frame-analysis.md contradicted itself (§5/§7 said the key was not
    recoverable, §0 says it is, verified 80/80) - §5/§7 marked superseded.
  * README status line, the physical-hardware checkbox and the 2W-handshake
    checkbox were stale - corrected to the real state.

Tests (225 total, +6):
  * TX UART framing is now asserted at the wire level (a dropped-framing
    regression previously passed every test); the packet-length test asserts
    literal sizes, not uart_wire_size() against itself.
  * The counter-wrap test crosses the real 2^32 boundary.
  * New negative pairing tests: a 1W node ignores pairing frames without
    crashing, a stranger's challenge mid-pairing is ignored, and the device is
    proven to have actually stored the adopted key.

Verified: 225 host tests (gcc ASan/UBSan and clang), gcc -Wconversion -Werror on
the changed sources, esphome compile of example.yaml and example-velux.yaml,
protocol-copy sync, and a heltec_wifi_lora_32_V2 firmware build.

+++ esphome/components/iown_homecontrol/iohome_2w.h
@@ -116,7 +116,11 @@ public:
 
 protected:
   ChannelState current_channel_;
-  unsigned long last_hop_time_us_;
+  // Stored as a fixed 32 bits so the wrap happens at 2^32 on every platform,
+  // matching Arduino micros() (a 32-bit unsigned long). On a 64-bit host,
+  // `unsigned long` is 64 bits and would only wrap at 2^64, so the timestamp
+  // math is done in uint32_t regardless of the caller's type.
+  uint32_t last_hop_time_us_;
   unsigned long hop_interval_us_;
   bool enabled_;
